+++ src/main/java/com/appman/nahug/repository/CatRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CatRepository extends JpaRepository<CatModel,Integer> {
+public interface CatRepository extends JpaRepository<CatModel,Long> {
 
 }
